Version.2
Dear Myself,
It's you, just with a little change in time.
I have padded the Header region and padded the Title and the image under the text too (you have to see if you can pad it using justify-content), didn't touch the footer or the text.
XOXO yourself, hope the site gets better!